Planet centric -> active pixel body
Update game.js to use the active pixel body (planet or serpent) rather
than always the planet.

Also update LittleGuy to attempt to have it work with the serpent as the
pixel body. This resulted in that only kinda working but also making
movement and updates more complicated and... buggier? Definitely worse.

There's also some half-pixel offset happening when rendering the serpent
and little guys.